+++ pipelinewise/fastsync/commons/tap_postgres.py
@@ -388,8 +388,8 @@ class FastSyncTapPostgres:
         if max_num:
             decimals = len(max_num.split('.')[1]) if '.' in max_num else 0
             decimal_format = f"""
-              'CASE WHEN "' || column_name || '" IS NULL THEN NULL ELSE GREATEST(LEAST({max_num}, ROUND("' || column_name || '"::numeric , {decimals})), -{max_num}) END'  # noqa E501
-            """
+              'CASE WHEN "' || column_name || '" IS NULL THEN NULL ELSE GREATEST(LEAST({max_num}, ROUND("' || column_name || '"::numeric , {decimals})), -{max_num}) END'
+            """ # noqa E501
             integer_format = """
               '"' || column_name || '"'
             """