+++ tests/end_to_end/test_target_snowflake.py
@@ -19,6 +19,7 @@ DIR = os.path.dirname(__file__)
 TAP_MARIADB_ID = 'mariadb_to_sf'
 TAP_MARIADB_SPLIT_LARGE_FILES_ID = 'mariadb_to_sf_split_large_files'
 TAP_MARIADB_BUFFERED_STREAM_ID = 'mariadb_to_sf_buffered_stream'
+TAP_MARIADB_REPLICA_ID = 'mariadb_replica_to_sf'
 TAP_POSTGRES_ID = 'postgres_to_sf'
 TAP_POSTGRES_SPLIT_LARGE_FILES_ID = 'postgres_to_sf_split_large_files'
 TAP_POSTGRES_ARCHIVE_LOAD_FILES_ID = 'postgres_to_sf_archive_load_files'
@@ -27,18 +28,18 @@ TAP_S3_CSV_ID = 's3_csv_to_sf'
 TARGET_ID = 'snowflake'
 
 
-# pylint: disable=attribute-defined-outside-init
+# pylint: disable=attribute-defined-outside-init,too-many-instance-attributes
 class TestTargetSnowflake:
     """
     End to end tests for Target Snowflake
     """
-
     def setup_method(self):
         """Initialise test project by generating YAML files from
         templates for all the configured connectors"""
 
         # Init query runner methods
         self.run_query_tap_mysql = self.e2e.run_query_tap_mysql
+        self.run_query_tap_mysql_2 = self.e2e.run_query_tap_mysql_2
         self.run_query_tap_postgres = self.e2e.run_query_tap_postgres
         self.run_query_target_snowflake = self.e2e.run_query_target_snowflake
         self.mongodb_con = self.e2e.get_tap_mongodb_connection()
@@ -57,14 +58,34 @@ class TestTargetSnowflake:
     def teardown_method(self):
         """Delete test directories and database objects"""
 
+    @pytest.fixture(autouse=True)
+    def skip_if_no_sf_credentials(self):
+        """
+        Test fixture to be used by all tests to check if SF is configured to decide whether to run the test or not.
+        """
+        # Skip every test if required env vars not provided
+        print('skip_if_no_sf_credentials executed')
+        if self.e2e.env['TARGET_SNOWFLAKE']['is_configured']:
+            yield
+        else:
+            pytest.skip('Target Snowflake environment variables are not provided')
+
+    @pytest.mark.dependency(name='validate')
+    def test_validate(self):
+        """Validate the YAML project with taps and target """
+
+        # validate project
+        return_code, stdout, stderr = tasks.run_command(f'pipelinewise validate --dir {self.project_dir}')
+
+        print('--------------- stdout ----------', stdout)
+        print('--------------- stderr ----------', stderr)
+        assert return_code == 0
+
+    @pytest.mark.dependency(depends=['validate'])
     @pytest.mark.dependency(name='import_config')
     def test_import_project(self):
         """Import the YAML project with taps and target and do discovery mode
         to write the JSON files for singer connectors"""
-
-        # Skip every target_postgres related test if required env vars not provided
-        if not self.e2e.env['TARGET_SNOWFLAKE']['is_configured']:
-            pytest.skip('Target Snowflake environment variables are not provided')
 
         # Setup and clean source and target databases
         self.e2e.setup_tap_mysql()
@@ -205,6 +226,25 @@ class TestTargetSnowflake:
             schema_postfix=self.snowflake_schema_postfix
         )
 
+    @pytest.mark.dependency(depends=['import_config'])
+    def test_replicate_mariadb_replica_to_sf(self):
+        """Replicate data from MariaDB to Snowflake"""
+        # 1. Run tap first time - both fastsync and a singer should be triggered
+        assertions.assert_run_tap_success(
+            TAP_MARIADB_REPLICA_ID, TARGET_ID, ['fastsync', 'singer']
+        )
+        assertions.assert_row_counts_equal(
+            self.run_query_tap_mysql_2,
+            self.run_query_target_snowflake,
+            self.snowflake_schema_postfix
+        )
+        assertions.assert_all_columns_exist(
+            self.run_query_tap_mysql_2,
+            self.e2e.run_query_target_snowflake,
+            mysql_to_snowflake.tap_type_to_target_type,
+            schema_postfix=self.snowflake_schema_postfix
+        )
+
     # pylint: disable=invalid-name
     @pytest.mark.dependency(depends=['import_config'])
     def test_resync_pg_to_sf_with_split_large_files(
@@ -335,9 +375,6 @@ class TestTargetSnowflake:
     @pytest.mark.dependency(depends=['import_config'])
     def test_replicate_pg_to_sf_with_archive_load_files(self):
         """Fastsync tables from Postgres to Snowflake with archive load files enabled"""
-        if not self.e2e.env['TARGET_SNOWFLAKE']['is_configured']:
-            pytest.skip('Target Snowflake environment variables are not provided')
-
         archive_s3_prefix = 'archive_folder'  # Custom s3 prefix defined in TAP_POSTGRES_ARCHIVE_LOAD_FILES_ID
 
         s3_bucket = os.environ.get('TARGET_SNOWFLAKE_S3_BUCKET')